Extract search-strategy-conditional
