+++ app/controllers/addresses_controller.rb
@@ -6,32 +6,11 @@
 #  https://github.com/hitobito/hitobito.
 
 class AddressesController < ApplicationController
+  include FullTextSearchStrategy
 
   skip_authorization_check
 
   def query
     render json: Address::FullTextSearch.new(query_param, search_strategy).typeahead_results
-  end
-
-  def query_param
-    params[:q]
-  end
-
-  private
-
-  def search_strategy
-    @search_strategy ||= search_strategy_class.new(current_user, query_param, params[:page])
-  end
-
-  def search_strategy_class
-    if sphinx?
-      SearchStrategies::Sphinx
-    else
-      SearchStrategies::Sql
-    end
-  end
-
-  def sphinx?
-    Hitobito::Application.sphinx_present?
   end
 end
